Widen before adding in rng_integer(): no signed overflow past 2^31
The Lemire offset can exceed INT_MAX when max - min >= 2^31, so
min + (int)offset overflowed int: undefined behaviour, reported by UBSan
on all three sanitizer legs for the golden case added in the previous
commit (zurand_engine.h:304, -2000000000 + -1498795907). Two's-complement
wraparound had been producing the right answer, since the sum lies in
[min, max], so output is unchanged; the sum is now formed in int64_t and
narrowed.

+++ src/zurand_engine.h
@@ -301,7 +301,9 @@ static void ZE_N(fill_integer_column)(int *out, R_xlen_t n, ZE_KEY_T key,
             if (!lemire_accept((uint32_t)buf[j], range, threshold, &offset))
                 offset = ZE_N(bounded_u32_retry)(key, (uint64_t)(w0 + j),
                                                  range, threshold);
-            o[j] = min + (int)offset;
+            /* offset can exceed INT_MAX when max - min >= 2^31; the sum
+             * itself lies in [min, max], so widen, add, then narrow. */
+            o[j] = (int)((int64_t)min + (int64_t)offset);
         }
     }
 }
